moving Printer to it's own class
